requirements.txt
flask==3.1.0
Werkzeug==3.1.4
requests==2.31.0

torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
tqdm==4.67.1
typing_extensions==4.15.0

numpy
pandas

opencv-python
Pillow

matplotlib
scikit-learn
gunicorn==21.2.0

